Semi-functional current graph, new page for graph widgets
